@@ -2,9 +2,9 @@ import 'package:webfeed/util/xml.dart';
 import 'package:xml/xml.dart';
 
 class AtomPerson {
-  final String name;
-  final String uri;
-  final String email;
+  final String? name;
+  final String? uri;
+  final String? email;
 
   AtomPerson({this.name, this.uri, this.email});
 
