@@ -2,9 +2,9 @@ import 'package:webfeed/util/xml.dart';
 import 'package:xml/xml.dart';
 
 class AtomSource {
-  final String id;
-  final String title;
-  final String updated;
+  final String? id;
+  final String? title;
+  final String? updated;
 
   AtomSource({
     this.id,
@@ -12,10 +12,8 @@ class AtomSource {
     this.updated,
   });
 
-  factory AtomSource.parse(XmlElement element) {
-    if (element == null) {
-      return null;
-    }
+  static parse(XmlElement? element) {
+    if (element == null) return null;
     return AtomSource(
       id: findFirstElement(element, 'id')?.text,
       title: findFirstElement(element, 'title')?.text,
