@@ -3,23 +3,23 @@ import 'package:webfeed/util/xml.dart';
 import 'package:xml/xml.dart';
 
 class DublinCore {
-  final String title;
-  final String description;
-  final String creator;
-  final String subject;
-  final String publisher;
-  final String contributor;
-  final DateTime date;
-  final DateTime created;
-  final DateTime modified;
-  final String type;
-  final String format;
-  final String identifier;
-  final String source;
-  final String language;
-  final String relation;
-  final String coverage;
-  final String rights;
+  final String? title;
+  final String? description;
+  final String? creator;
+  final String? subject;
+  final String? publisher;
+  final String? contributor;
+  final DateTime? date;
+  final DateTime? created;
+  final DateTime? modified;
+  final String? type;
+  final String? format;
+  final String? identifier;
+  final String? source;
+  final String? language;
+  final String? relation;
+  final String? coverage;
+  final String? rights;
 
   DublinCore({
     this.title,
@@ -41,7 +41,7 @@ class DublinCore {
     this.rights,
   });
 
-  factory DublinCore.parse(XmlElement element) {
+  static parse(XmlElement? element) {
     if (element == null) {
       return null;
     }
