@@ -5,9 +5,9 @@ import 'package:webfeed/util/xml.dart';
 import 'package:xml/xml.dart';
 
 class Community {
-  final StarRating starRating;
-  final Statistics statistics;
-  final Tags tags;
+  final StarRating? starRating;
+  final Statistics? statistics;
+  final Tags? tags;
 
   Community({
     this.starRating,
@@ -15,10 +15,8 @@ class Community {
     this.tags,
   });
 
-  factory Community.parse(XmlElement element) {
-    if (element == null) {
-      return null;
-    }
+  static parse(XmlElement? element) {
+    if (element == null) return null;
     return Community(
       starRating: StarRating.parse(
         findFirstElement(element, 'media:starRating'),
