@@ -1,20 +1,20 @@
 import 'package:xml/xml.dart';
 
 class Content {
-  final String url;
-  final String type;
-  final int fileSize;
-  final String medium;
-  final bool isDefault;
-  final String expression;
-  final int bitrate;
-  final double framerate;
-  final double samplingrate;
-  final int channels;
-  final int duration;
-  final int height;
-  final int width;
-  final String lang;
+  final String? url;
+  final String? type;
+  final int? fileSize;
+  final String? medium;
+  final bool? isDefault;
+  final String? expression;
+  final int? bitrate;
+  final double? framerate;
+  final double? samplingrate;
+  final int? channels;
+  final int? duration;
+  final int? height;
+  final int? width;
+  final String? lang;
 
   Content({
     this.url,
