@@ -1,18 +1,16 @@
 import 'package:xml/xml.dart';
 
 class Param {
-  final String name;
-  final String value;
+  final String? name;
+  final String? value;
 
   Param({
     this.name,
     this.value,
   });
 
-  factory Param.parse(XmlElement element) {
-    if (element == null) {
-      return null;
-    }
+  static Param? parse(XmlElement? element) {
+    if (element == null) return null;
     return Param(
       name: element.getAttribute('name'),
       value: element.text,
