@@ -1,9 +1,9 @@
 import 'package:xml/xml.dart';
 
 class PeerLink {
-  final String type;
-  final String href;
-  final String value;
+  final String? type;
+  final String? href;
+  final String? value;
 
   PeerLink({
     this.type,
@@ -11,10 +11,8 @@ class PeerLink {
     this.value,
   });
 
-  factory PeerLink.parse(XmlElement element) {
-    if (element == null) {
-      return null;
-    }
+  static parse(XmlElement? element) {
+    if (element == null) return null;
     return PeerLink(
       type: element.getAttribute('type'),
       href: element.getAttribute('href'),
