@@ -1,9 +1,9 @@
 import 'package:xml/xml.dart';
 
 class Restriction {
-  final String relationship;
-  final String type;
-  final String value;
+  final String? relationship;
+  final String? type;
+  final String? value;
 
   Restriction({
     this.relationship,
@@ -11,10 +11,8 @@ class Restriction {
     this.value,
   });
 
-  factory Restriction.parse(XmlElement element) {
-    if (element == null) {
-      return null;
-    }
+  static parse(XmlElement? element) {
+    if (element == null) return null;
     return Restriction(
       relationship: element.getAttribute('relationship'),
       type: element.getAttribute('type'),
