@@ -2,10 +2,10 @@ import 'package:webfeed/util/xml.dart';
 import 'package:xml/xml.dart';
 
 class Scene {
-  final String title;
-  final String description;
-  final String startTime;
-  final String endTime;
+  final String? title;
+  final String? description;
+  final String? startTime;
+  final String? endTime;
 
   Scene({
     this.title,
@@ -14,10 +14,8 @@ class Scene {
     this.endTime,
   });
 
-  factory Scene.parse(XmlElement element) {
-    if (element == null) {
-      return null;
-    }
+  static Scene? parse(XmlElement? element) {
+    if (element == null) return null;
     return Scene(
       title: findFirstElement(element, 'sceneTitle')?.text,
       description: findFirstElement(element, 'sceneDescription')?.text,
