@@ -1,10 +1,10 @@
 import 'package:xml/xml.dart';
 
 class StarRating {
-  final double average;
-  final int count;
-  final int min;
-  final int max;
+  final double? average;
+  final int? count;
+  final int? min;
+  final int? max;
 
   StarRating({
     this.average,
@@ -13,12 +13,12 @@ class StarRating {
     this.max,
   });
 
-  factory StarRating.parse(XmlElement element) {
+  factory StarRating.parse(XmlElement? element) {
     return StarRating(
-      average: double.tryParse(element.getAttribute('average') ?? '0'),
-      count: int.tryParse(element.getAttribute('count') ?? '0'),
-      min: int.tryParse(element.getAttribute('min') ?? '0'),
-      max: int.tryParse(element.getAttribute('max') ?? '0'),
+      average: double.tryParse(element?.getAttribute('average') ?? '0'),
+      count: int.tryParse(element?.getAttribute('count') ?? '0'),
+      min: int.tryParse(element?.getAttribute('min') ?? '0'),
+      max: int.tryParse(element?.getAttribute('max') ?? '0'),
     );
   }
 }
