@@ -1,18 +1,18 @@
 import 'package:xml/xml.dart';
 
 class Statistics {
-  final int views;
-  final int favorites;
+  final int? views;
+  final int? favorites;
 
   Statistics({
     this.views,
     this.favorites,
   });
 
-  factory Statistics.parse(XmlElement element) {
+  factory Statistics.parse(XmlElement? element) {
     return Statistics(
-      views: int.tryParse(element.getAttribute('views') ?? '0'),
-      favorites: int.tryParse(element.getAttribute('favorites') ?? '0'),
+      views: int.tryParse(element?.getAttribute('views') ?? '0'),
+      favorites: int.tryParse(element?.getAttribute('favorites') ?? '0'),
     );
   }
 }
