@@ -5,13 +5,14 @@ void main() async {
   var client = http.Client();
 
   // RSS feed
-  var response = await client
-      .get('https://developer.apple.com/news/releases/rss/releases.rss');
+  var response = await client.get(
+      Uri(path: 'https://developer.apple.com/news/releases/rss/releases.rss'));
   var channel = RssFeed.parse(response.body);
   print(channel);
 
   // Atom feed
-  response = await client.get('https://www.theverge.com/rss/index.xml');
+  response =
+      await client.get(Uri(path: 'https://www.theverge.com/rss/index.xml'));
   var feed = AtomFeed.parse(response.body);
   print(feed);
 
