@@ -2,7 +2,7 @@ import 'package:xml/xml.dart';
 
 enum ItunesEpisodeType { full, trailer, bonus, unknown }
 
-ItunesEpisodeType newItunesEpisodeType(XmlElement element) {
+ItunesEpisodeType newItunesEpisodeType(XmlElement? element) {
   switch (element?.text) {
     case 'full':
       return ItunesEpisodeType.full;
