@@ -2,7 +2,7 @@ import 'package:xml/xml.dart';
 
 enum ItunesType { episodic, serial, unknown }
 
-ItunesType newItunesType(XmlElement element) {
+ItunesType newItunesType(XmlElement? element) {
   switch (element?.text) {
     case 'episodic':
       return ItunesType.episodic;
