@@ -12,16 +12,16 @@ final _imagesRegExp = RegExp(
 ///
 class RssContent {
   String value;
-  Iterable<String> images;
+  Iterable<String?> images;
 
   RssContent(this.value, this.images);
 
-  factory RssContent.parse(XmlElement element) {
-    if (element == null) {
+  static parse(XmlElement? element) {
+    if(element == null) {
       return null;
     }
-    final content = element.text;
-    final images = <String>[];
+    final dynamic? content = element.text;
+    final images = <String?>[];
     _imagesRegExp.allMatches(content).forEach((match) {
       images.add(match.group(1));
     });
