@@ -2,11 +2,11 @@ import 'dart:core';
 
 import 'package:xml/xml.dart';
 
-XmlElement findFirstElement(
-  XmlNode node,
+XmlElement? findFirstElement(
+  XmlNode? node,
   String name, {
   bool recursive = false,
-  String namespace,
+  String? namespace,
 }) {
   try {
     return findElements(node, name, recursive: recursive, namespace: namespace)
@@ -16,17 +16,17 @@ XmlElement findFirstElement(
   }
 }
 
-Iterable<XmlElement> findElements(
-  XmlNode node,
+Iterable<XmlElement>? findElements(
+  XmlNode? node,
   String name, {
   bool recursive = false,
-  String namespace,
+  String? namespace,
 }) {
   try {
     if (recursive) {
-      return node.findAllElements(name, namespace: namespace);
+      return node?.findAllElements(name, namespace: namespace);
     } else {
-      return node.findElements(name, namespace: namespace);
+      return node?.findElements(name, namespace: namespace);
     }
   } on StateError {
     return null;
@@ -34,7 +34,7 @@ Iterable<XmlElement> findElements(
 }
 
 bool parseBoolLiteral(XmlElement element, String tagName) {
-  var v = findFirstElement(element, tagName)?.text?.toLowerCase()?.trim();
+  var v = findFirstElement(element, tagName)?.text.toLowerCase().trim();
   if (v == null) return false;
   return ['yes', 'true'].contains(v);
 }
